chore: refactor MLPClassifier and MLPRegressor to use centralized activation functions and derivatives from utils.py
